--- a/Conch/cpanfile
+++ b/Conch/cpanfile
@@ -12,6 +12,7 @@ requires 'aliased';
 requires 'Try::Tiny';
 requires 'Class::StrongSingleton';
 requires 'Time::HiRes';
+requires 'Time::Moment';
 
 # mojolicious
 requires 'Mojolicious';
@@ -71,4 +72,5 @@ on 'test' => sub {
     requires 'IO::All';
     requires 'JSON::Validator';
     requires 'YAML::XS';
+	requires 'Test::Exception';
 };