--- a/Conch/lib/Conch/Time.pm
+++ b/Conch/lib/Conch/Time.pm
@@ -32,8 +32,11 @@ use overload
 	eq   => 'compare',
 	ne   => sub { !compare(@_) };
 
-use constant PG_TIMESTAMP_FORMAT =>
-	qr/^(\d{4,})-(\d{2,})-(\d{2,}) (\d{2,}):(\d{2,}):(\d{2,})(\.\d+)?([-\+][\d:]+)$/;
+use constant PG_TIMESTAMP_FORMAT => qr/
+	^(\d{4,})-(\d{2,})-(\d{2,})\s
+	(\d{2,}):(\d{2,}):(\d{2,})(\.\d+)
+	?([-\+][\d:]+)$
+/x;
 
 =head2 timestamp