--- a/Conch/lib/Conch/Time.pm
+++ b/Conch/lib/Conch/Time.pm
@@ -11,7 +11,6 @@ Conch::Time - format Postgres Timestamps as RFC 3337 UTC timestamps
 	my $postgres_timestamp = '2018-01-26 12:24:18.893874-07';
 	my $time = Conch::Time->new($postgres_timestamp);
 
-	say $time; # '2018-01-26T12:24:18.893Z'
 	$time eq $time; # 1
 
 
@@ -23,46 +22,61 @@ package Conch::Time;
 use Mojo::Base -base, -signatures;
 
 use POSIX qw(strftime);
+use Time::Moment;
 use Time::HiRes;
-use DateTime::Format::Strptime;
 use Mojo::Exception;
 
 use overload
-	'""' => 'to_string',
+	'""' => 'rfc3339',
 	eq   => 'compare',
 	ne   => sub { !compare(@_) };
 
 use constant PG_TIMESTAMP_FORMAT => qr/
 	^(\d{4,})-(\d{2,})-(\d{2,})\s
-	(\d{2,}):(\d{2,}):(\d{2,})(\.\d+)
-	?([-\+][\d:]+)$
+	(\d{2,}):(\d{2,}):(\d{2,})\.?(\d+)
+	?([-\+])([\d:]+)$
 /x;
 
-=head2 timestamp
 
-Underlying RFC 3339 formatted timestamp
 
-=cut
-
-has 'timestamp';
+has 'moment';
 
 =head2 new
+
+	Conch::Time->new($pg_timestamptz);
+
 =cut
 
-
 sub new ( $class, $timestamptz ) {
+	my @c = ( $timestamptz =~ PG_TIMESTAMP_FORMAT );
 	Mojo::Exception->throw('Invalid Postgres timestamp')
-		unless $timestamptz && ( $timestamptz =~ m/${\PG_TIMESTAMP_FORMAT}/ );
-	my $dt = "$1-$2-$3T$4:$5:$6." . _normalize_millisec($7) . _normalize_tz($8);
-	$class->SUPER::new( timestamp => $dt );
+		unless @c;
+
+	$c[6] = 0 unless $c[6];
+
+	my $off_minutes;
+	if ($c[8] =~ /:/) {
+		my ($hours, $minutes) = $c[8] =~ /^(\d\d)[:]?(\d\d)$/;
+
+		$off_minutes = ($hours*60);
+		$off_minutes = $off_minutes + $minutes if $minutes;
+	} else {
+		$off_minutes = $c[8] * 60;
+	}
+
+	my $m = Time::Moment->new(
+		year       => $c[0],
+		month      => $c[1],
+		day        => $c[2],
+		hour       => $c[3],
+		minute     => $c[4],
+		second     => $c[5],
+		nanosecond => $c[6]*1000,
+		offset     => "${c[7]}${off_minutes}",
+	);
+	return $class->SUPER::new(moment => $m);
 }
 
-sub _from_hires($class, $epoch, $mil) {
-	my $dt = strftime("%Y-%m-%dT%H:%M:%S", gmtime($epoch)) .
-		_normalize_millisec($mil) . "Z";
-
-	return $class->SUPER::new(timestamp => $dt);
-}
 
 
 =head2 now
@@ -72,46 +86,29 @@ sub _from_hires($class, $epoch, $mil) {
 Return an object based on the current time.
 
 Time are high resolution and will generate unique timestamps to the
-millisecond.
+nanosecond.
 
 =cut
 
 sub now ($class) {
-	return $class->_from_hires(Time::HiRes::gettimeofday());
+	return $class->SUPER::new(moment => Time::Moment->now());
 }
 
-# Given a float, return the number of integer milliseconds it represents
-sub _normalize_millisec {
-	substr( sprintf( '%.3f', shift || 0 ), 2 );
-}
+=head2 from_epoch
 
-sub _normalize_tz {
-	my $tz = shift;
-	# return 'Z' if the timezone is 00 or 00:00
-	return 'Z' if $tz =~ /^[-\+]00(?!:[1-9]\d)/;
-	# Append :00 if the timezone doesn't specify minutes
-	return $tz . ':00' if $tz =~ /^[-\+]\d\d$/;
+	Conch::Time->from_epoch(time());
 
-	# Munge offsets like -0500 into -05:00
-	return "$1$2:$3" if $tz =~ /^([-\+])(\d\d)(\d\d)$/;
-
-	return $tz;
-}
-
-=head2 to_datetime
-
-Return a C<DateTime> object representing the timestamp.
-
-B<NOTE:> This method will negatively impact performance if called frequently.
+	Conch::Time->from_epoch(Time::HiRes::gettimeofday);
 
 =cut
 
-sub to_datetime {
-	return DateTime::Format::Strptime->new(
-		pattern  => '%Y-%m-%dT%H:%M:%S.%3N%z',
-		on_error => 'croak'
-	)->parse_datetime( shift->timestamp );
+sub from_epoch ($class, $epoch, $nano = 0) {
+	return $class->SUPER::new(moment => Time::Moment->from_epoch(
+		$epoch,
+		$nano,
+	));
 }
+
 
 =head2 compare
 
@@ -121,19 +118,44 @@ Compare two Conch::Time objects. Used to overload C<eq> and C<ne>.
 
 sub compare {
 	my ( $self, $other ) = @_;
-	$self->timestamp eq $other->timestamp;
+	return $self->moment->is_equal($other->moment)
 }
 
-=head2 to_string
 
-Render the timestamp as a RFC 3337 timestamp string. Used to
+=head2 CONVERSIONS
+
+=head3 rfc3339
+
+Return an RFC3339 compatible string
+
+=cut
+
+sub rfc3339 {
+	my $self = shift;
+	return $self->moment->strftime("%Y-%m-%dT%H:%M:%S.%3N%Z");
+}
+
+
+
+
+=head3 timestamp
+
+Return an RFC3339 compatible string
+
+=cut
+
+sub timestamp { shift->rfc3339() }
+
+
+
+=head3 to_string
+
+Render the timestamp as a RFC 3339 timestamp string. Used to
 overload string coercion.
 
 =cut
 
-sub to_string {
-	shift->timestamp;
-}
+sub to_string { shift->rfc3339 }
 
 1;