Using Time::Moment internally lets us perform date/time calculations and render various string outputs   without introducing the slowdowns we were seeing with DateTime objects.

Brings in Time::Moment and Test::Exception as new dependencies

_wrk results:_

Before:
```
wrk -H "Cookie: conch=$COOKIE" --duration 60 --latency -T 20 http://localhost:5001/workspace/9d374081-ba10-496e-88ed-c1fa2a51a7c2/device                                                                                                     
Running 1m test @ http://localhost:5001/workspace/9d374081-ba10-496e-88ed-c1fa2a51a7c2/device                                                                                                                                                    
  2 threads and 10 connections                                                                                                                                                                                                                   
  Thread Stats   Avg      Stdev     Max   +/- Stdev                                                                                                                                                                                              
    Latency     1.46s    44.38ms   1.63s    78.05%                                                                                                                                                                                               
    Req/Sec     2.98      0.16     3.00     97.56%                                                                                                                                                                                               
  Latency Distribution                                                                                                                                                                                                                           
     50%    1.45s                                                                                                                                                                                                                                
     75%    1.49s                                                                                                                                                                                                                                
     90%    1.51s                                                                                                                                                                                                                                
     99%    1.63s                                                                                                                                                                                                                                
  410 requests in 1.00m, 205.46MB read                                                                                                                                                                                                           
Requests/sec:      6.83                                                                                                                                                                                                                          
Transfer/sec:      3.42MB     
```                                                                                                                                                                                                                   
                                                                                                                                                                                                                                                 
After:  
```
$ wrk -H "Cookie: conch=$COOKIE" --duration 60 --latency -T 20 http://localhost:5001/workspace/9d374081-ba10-496e-88ed-c1fa2a51a7c2/device                                                                                                     
Running 1m test @ http://localhost:5001/workspace/9d374081-ba10-496e-88ed-c1fa2a51a7c2/device                                                                                                                                                    
  2 threads and 10 connections                                                                                                                                                                                                                   
  Thread Stats   Avg      Stdev     Max   +/- Stdev                                                                                                                                                                                              
    Latency     1.36s    34.29ms   1.46s    70.45%                                                                                                                                                                                               
    Req/Sec     3.05      0.21     4.00     95.45%                                                                                                                                                                                               
  Latency Distribution                                                                                                                                                                                                                           
     50%    1.36s                                                                                                                                                                                                                                
     75%    1.39s                                                                                                                                                                                                                                
     90%    1.40s                                                                                                                                                                                                                                
     99%    1.46s                                                                                                                                                                                                                                
  440 requests in 1.00m, 220.49MB read                                                                                                                                                                                                           
Requests/sec:      7.33                                                                                                                                                                                                                          
Transfer/sec:      3.67MB           
```